FDG-5310 Moved about pages
