+++ src/components/contact-section/contact-section.module.scss
@@ -1,0 +1,5 @@
+@use '../../variables.module' as *;
+
+.mainSection {
+  overflow: hidden;
+}
