--- src/pages/about-us/contact-section/contact-section.module.scss
@@ -1,5 +1,0 @@
-@use '../../../variables.module.scss' as *;
-
-.mainSection {
-  overflow: hidden;
-}
